slack_bot/requirements.txt
slackclient
SQLAlchemy
psycopg2-binary
vaderSentiment
pymongo
